refactor: refactoring session context into tsx files
