--- front/src/hooks/useSession.js
@@ -1,8 +1,0 @@
-import { useContext } from "react"
-import SessionContext from "../context/SessionProvider"
-
-const useSession = () => {
-  return useContext(SessionContext);
-}
-
-export default useSession;
